Basic structure for graphs created
An interface including the main behaviour of a graph was included in .

In , a constructor with some basic methods were created for a struct representing a graph, modelled by its adjacency matrix.   An initial decision is not to calculate the degree sequence until the first time it is required.

Some basic tests were also added for .