feat: add constructor for complete graphs
A constructor was created for complete graphs modelled by their adjacency matrix.  A function to verify whether a graph is complete was also provided.

The constructor simply creates a matrix with every non-diagonal entry equal to 1 (and diagonal equal to 0).   The function to verify completeness checks the adjacency matrix to have the same property.