test: add test to constructor for complete graphs
Adds a test for the complete graph generator, for graphs modelled by their adjacency matrix.    The test checks whether every non-diagonal entry of the adjacency matrix of the graph created by CompleteMatrixGraph is equal to 1, and every diagonal entry is equal to 0.

A test is also added for the IsCompleteMatrixGraph function. The test is performed by hardcoding both a complete graph, and graphs that fail to be complete, either by having a diagonal 1, or a non-diagonal 0.