Adds digraphs modelled by adjacency matrices
A way to differentiate digraphs with adjacency matrices from
regular graphs was needed for implementation of algorithms in
the future.

A file adjdigraph.go with the MatrixDigraph struct and some of
its functionality was added. Functionality added in this commit:
- Function to obtain adjacency matrix.
- Function to obtain order of digraph.
- Functions to obtain degree, indegree, and outdegree sequences.
A testing file was also added to test the functions Adjacency,
Order, and DegreeSequence.