feat: adds a constructor and a recognizer for cycles
Adds a constructor for irreflexive cycle graphs.  The constructor generates a cycle in its canonical circular order, i.e., vertex `i` is adjacent to vertices `i-1` and `i+1` `mod n`.   The constructor is `CycleMatrixGraph`.

Also adds the function `IsCycleMatrixGraph`, to recognize cycles.  This function recognizes an irreflexive cycle disregarding the ordering of its vertices. It might be considered as an isomorphism test for cycles.