test: adds tests for the cycle generator
Adds test for both the cycle generator and the cycle recognizer.

For the cycle generator, it is checked that the adjacency matrix of the graph generated by calling `CycleMatrixGraph` corresponds to the adjacency matrix of an irreflexive cycle with the canonical circular ordering of its vertices.   Three cycles of random lengths between 3 and 1000 are constructed.

For the cycle recognizer, the behaviour of the function `IsCycleMatrixGraph` is verified against hardcoded examples of cycle and non-cycle graphs.