feat: Adds constructor for circulant digraphs
A constructor for circulant digraphs is added, such that it receives an integer (the number of vertices) and a `map[int]bool` (the jumps) to create the adjacency matrix of a `MatrixDigraph`.

This constructor handles both positive and negative jumps, and **ignores jumps of length with absolute value at least the number of vertices**.

Maybe it might be a good idea to use this constructor to obtain the undirected graph analogue.   Otherwise, an undirected graph analogue should be written, possibly reusing some code from this function.