test: for the CirculantMatrixDigraph constructor
A test is added for the CirculantMatrixDigraph function, randomly creating five sets of jumps, together with the corresponding adjacency matrix.  Then, the CirculantMatrixDigraph is called with this map of jumps, and the adjacency matrix of the resulting digraph is compared to the ones previously constructed.

A final test hardcodes an adjacency matrix with both positive and negative jumps, and compares it with the adjacency matrix of the digraph obtained by calling CirculantMatrixDigraph.