feat: adds constructor and recognizer for paths
Adds a constructor for irreflexive path graphs. The constructor generates a path
in its canonical linear order, i.e., vertex `i` is adjacent to vertices `i-1` and
`i+1` `mod n`, except for the first vertex, which is only adjacent to the second,
and the last vertex, which is only adjacent to its immediate predecessor. The
constructor is `PathMatrixGraph`.

Also adds the function `IsPathMatrixGraph`, to recognize paths. This function
recognizes an irreflexive path disregarding the ordering of its vertices. It might
be considered as an isomorphism test for paths.