fix: fixes problems pointed out in PR #20
Besides correcting the problems in the comments, tests where added to check
whether a trivial graph is a complete graph, and whether a trivial graph is
a path graph.   It was necessary to modify the recognizer for path graphs.

Also, now the constructor for cycles uses the constructor for paths.  A path
is generated, and the edge to close the cycle is then added.