refactor: changes MatrixGraph for Digraph structure
This is a work in progress. Graphs and digraphs should be
unified. Also, both representations of these structures
must be included in the general `Digraph` struct. Two
booleans are added to `Digraph`, in order to indicate
whether it is a graph, and if the adjacency list is
up-to-date with the adjacency matrix after adding or
deleting edges.

Refactoring in all classes is needed.