feat: Added ByteMatrixToIntSlice function to sliceutils.
If a byte slice represents a binary number (e.g. (1, 1, 0, 0, 1, 1)) then the desired
integer would be 35, however all of the currently available standart libraries for
byte and bit operations would consider each bit as a byte, thus the conversion would
differ from 35. To solve this, we had to implement it manually, the function that does
this is ByteSliceToInt.