fix: Graph6 appending zero's problem.
Given a vector x of bits. The length of x must be a multiple of 6
to obtain a graph6 format, if length is not a multiple of 6 then
we append zeros. If the vector x corresponds to the edges of the
graph, append zeros to the right. Otherwise, the vector x.
corresponds to the order of the graph, append zeros to the left.
The previous implementation of ToGraph6 always appended zeros to
the left, this was fixed by refactoring the function ExtendByteSlice
to indicate in which direction the zero's will be appended.