feat: Added a Block Map structure.
Go doesn't support map that take an structure as a key. To get around this the
Block Map structure is implemented. The methods are somewhat limited as we only
need addition and contention operataions. The hash function is just the sum of
the rows and columns indexes. It is relatively simple and as sets do not have an
order then same sets with its element on different order yield the same hash
value. The only drawback is the amount of collisions we can have; this is only a
temporal solution and can later be changed if a better hash function is found.