feat: updateAffectedBlocksColumns function implemented.
Given a block B, and a column refinement C', C'' of B. The previous refinement
not only affects B but all of the blocks below B. The function updates the size
of all of these blocks using only the smallest partition in the
refinement. This, usually, speeds up a lot the process of calculating the size
of all the new produced blocks.