feat: updateAffectedBlocksRows function implemented.
Similarly to column refinements, given a block B and a row refinement R', R'' of
B. The previous refinement not only affects B but all of the blocks next to B
too. The function updates the size of all these blocks using only the smallest
partition in the refinement. This, usually, speeds up a lot the process of
calculating the size of all the new produced blocks.