feat: Unit tests for the Double Lexicographical Ordering Algorithm.
The unit test takes several arbitrary matrices of varying sizes and tests that
the algorithm returns an ordered output several time. The probability of failure
after certain amount of testing goes down exponentially. If the algorithm hasn't
returned an error after a certain amount of tries, there is a "really good"
chance that the algorithm was correctly implemented.